+++ packages/dumb/src/components/trends/cbg/cbg-median-animated/cbg-median-animated.tsx
@@ -29,10 +29,10 @@ import React, { FunctionComponent } from 'react'
 import { spring, TransitionMotion } from '@serprex/react-motion'
 
 import styles from './cbg-median-animated.css'
-import { getBgClass } from '../../../utils/blood-glucose/blood-glucose.util'
-import { BgBounds, ClassificationType } from '../../../models/blood-glucose.model'
-import { springConfig } from '../../../models/constants/animation.constants'
-import { CbgMedianTransitionMotionInterpolate } from '../../../models/animation.model'
+import { getBgClass } from '../../../../utils/blood-glucose/blood-glucose.util'
+import { BgBounds, ClassificationType } from '../../../../models/blood-glucose.model'
+import { springConfig } from '../../../../models/constants/animation.constants'
+import { CbgMedianTransitionMotionInterpolate } from '../../../../models/animation.model'
 
 interface CbgMedianAnimatedProps {
   bgBounds: BgBounds